+++ openchpl_views.sql
@@ -16,7 +16,7 @@ LEFT JOIN (SELECT certification_criterion_id, number, title FROM openchpl.certif
 
 ON a.certification_criterion_id = b.certification_criterion_id;
 
-ALTER VIEW openchpl.certification_result_details OWNER TO openchpl;
+-- ALTER VIEW openchpl.certification_result_details OWNER TO openchpl;
 
 CREATE OR REPLACE VIEW openchpl.cqm_result_details AS
 
@@ -44,7 +44,7 @@ LEFT JOIN openchpl.cqm_criterion b ON a.cqm_criterion_id = b.cqm_criterion_id
 
 LEFT JOIN openchpl.cqm_version c ON b.cqm_version_id = c.cqm_version_id;
 
-ALTER VIEW openchpl.cqm_result_details OWNER TO openchpl;
+-- ALTER VIEW openchpl.cqm_result_details OWNER TO openchpl;
 
 CREATE OR REPLACE VIEW openchpl.certified_product_details AS
 
@@ -57,7 +57,6 @@ a.testing_lab_id,
 a.certification_body_id,
 a.chpl_product_number,
 a.report_file_location,
-a.quality_management_system_att,
 a.acb_certification_id,
 a.practice_type_id,
 a.product_classification_type_id,
@@ -73,6 +72,9 @@ a.certified_date_code,
 a.visible_on_chpl,
 a.terms_of_use_url,
 a.api_documentation_url,
+a.ics,
+a.sed,
+a.qms,
 b.year,
 c.certification_body_name,
 c.certification_body_code,
@@ -90,7 +92,9 @@ COALESCE(m.count_cqms, 0) as "count_cqms",
 COALESCE(o.count_corrective_action_plans, 0) as "count_corrective_action_plans",
 a.last_modified_date,
 n.certification_status_name,
-p.transparency_attestation
+p.transparency_attestation,
+q.testing_lab_name,
+q.testing_lab_code
 
 FROM openchpl.certified_product a
 
@@ -120,6 +124,8 @@ LEFT JOIN (SELECT certified_product_id, count(*) as "count_certifications" FROM 
 LEFT JOIN (SELECT certified_product_id, count(*) as "count_cqms" FROM (SELECT DISTINCT ON (cqm_id, certified_product_id) * FROM openchpl.cqm_result_details WHERE success = true AND deleted <> true) l GROUP BY certified_product_id ORDER BY certified_product_id) m ON a.certified_product_id = m.certified_product_id
 
 LEFT JOIN (SELECT certified_product_id, count(*) as "count_corrective_action_plans" FROM (SELECT * FROM openchpl.corrective_action_plan WHERE deleted <> true) n GROUP BY certified_product_id) o ON a.certified_product_id = o.certified_product_id
+
+LEFT JOIN (SELECT testing_lab_id, name as "testing_lab_name", testing_lab_code from openchpl.testing_lab) q on a.testing_lab_id = q.testing_lab_id
 ;
 
-ALTER VIEW openchpl.certified_product_details OWNER TO openchpl;
+-- ALTER VIEW openchpl.certified_product_details OWNER TO openchpl;